refactor(triple-fallback-action): split files
