+++ packages/triple-fallback-action/src/constant.tsx
@@ -1,0 +1,1 @@
+export const TRIPLE_FALLBACK_ACTION_CLASS_NAME = '-triple-fallback-action'
