+++ packages/triple-fallback-action/src/index.ts
@@ -1,1 +1,4 @@
+export * from './constant'
+export * from './triple-fallback-action-remover'
 export * from './triple-fallback-action'
+export * from './use-triple-fallback-action-remover'
