+++ packages/triple-fallback-action/src/triple-fallback-action-remover.tsx
@@ -1,0 +1,7 @@
+import { useTripleFallbackActionRemover } from './use-triple-fallback-action-remover'
+
+export function TripleFallbackActionRemover() {
+  useTripleFallbackActionRemover()
+
+  return null
+}
